fetching users to frontend with query
