+++ Client/src/App.css
@@ -11,7 +11,8 @@ box-sizing: border-box;
   text-align: center;
   display: flex;
   flex-direction: column;
+  align-items: center;
   justify-content: flex-start;
-  background: #333;
-  color: #535bf2;
+  background: #fff;
+  color: #333;
 }
