+++ Client/src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css'
 import Header from "./Components/Header.tsx";
 import { ApolloProvider, ApolloClient, InMemoryCache } from "@apollo/client";
-import Users from "./Components/Users.tsx";
+import Users from "./Components/User/Users.tsx";
 
 const client = new ApolloClient({
 uri: "http://localhost:5000/graphql",
