+++ Client/src/Components/User/UserRow.scss
@@ -1,0 +1,10 @@
+.user-row{
+  width: 30rem;
+  height: 3rem;
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  background: aliceblue;
+  padding: 0 20px;
+  margin-bottom: 10px;
+}
