+++ Client/src/queries/userQueries.ts
@@ -1,0 +1,14 @@
+import {gql} from "@apollo/client";
+
+const GET_USERS = gql`
+    query getUsers {
+    users{
+        id
+        name
+        email
+        phone
+       } 
+    }
+    `;
+
+export {GET_USERS};
